Adding working chunks
Chunks work!! Now you can initialize a chunk from voxel.h, and arrange them
in order. They produce meshes for you that you can integrate into the
render queue, and they work quite solidly. I've tested a majority of the
features.

However, we could do for some upgrading on the textures. I plan to be
able to use a numerous amount of textures and seeing as we wont be giving
the user the responsibility of a list of meshes, I think we should get
to work on making a texture-atlas system or we could wrangle our meshes
to get a mesh variant that can bind multiple texture objects per vertex
and coordinate textures for certain vertices? Maybe not.

But regardless, the demo works. I also was able to add in some fixes to
the perspective for the camera. It sorta looks a bit better for it since
it doesnt warp by the value, but I think I will have to fix this in the
future.

WAWACRAFT:EVOLVED FOREVER!!!

And seeing as this is valentine's day ;), I think I should push my commits
to my most lovable project ever: Wawacraft:Evolved.

+++ src/mesh.cpp
@@ -244,7 +244,7 @@ void camera::render_mesh(mesh_3d* mesh)
   mesh->render();
 }
 
-void camera::resolution_set(int px, int py)
+void camera::resolution_set(float px, float py)
 {
   c_resolution_x=px;
   c_resolution_y=py;
